Make some improvements to the Python script.
- Don't allow CTRL+C while a Notecard transaction is ongoing. Previously, if the
user interrupted the script with CTRL+C during the middle of a transaction and
immediately ran the script again, it was likely for the old transaction to still
be coming over the I2C bus. This wrecks communication between the script and the
Notecard. Instead, note that CTRL+C was issued, finish the current transaction,
and then interrupt the script.
- Retry web.get requests if the status code isn't 206 or 200. I saw a 503 error
at one point during testing. This change makes the script resilient to these
kinds of errors.
- Format everything with YAPF.